+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [10.0.0-alpha.9](https://github.com/videojs/v10/compare/@videojs/icons@10.0.0-alpha.8...@videojs/icons@10.0.0-alpha.9) (2026-03-06)
+
+
+### Miscellaneous Chores
+
+* **@videojs/icons:** Synchronize videojs versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @videojs/utils bumped to 10.0.0-alpha.9
+
+## [10.0.0-alpha.8](https://github.com/videojs/v10/compare/@videojs/icons@10.0.0-alpha.7...@videojs/icons@10.0.0-alpha.8) (2026-03-05)
+
+
+### Miscellaneous Chores
+
+* **@videojs/icons:** Synchronize videojs versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @videojs/utils bumped to 10.0.0-alpha.8
+
 ## [10.0.0-alpha.7](https://github.com/videojs/v10/compare/@videojs/icons@10.0.0-alpha.6...@videojs/icons@10.0.0-alpha.7) (2026-03-04)